fix: 채팅방 조회 참여일 기준으로 변경, 마이 페이지 및 상세 조회 인원 계산 수정

+++ src/main/java/com/team573/gongguri/domain/groupPurchase/dto/GroupPurchaseDetailResponseDto.java
@@ -9,7 +9,7 @@ public record GroupPurchaseDetailResponseDto(
         String content,
         int price,
         int maxParticipants,
-        int currentParticipants,
+        Long currentParticipants,
         String bank,
         String account,
         String progressStatus,
