+++ src/main/java/com/team573/gongguri/domain/groupPurchase/dto/GroupPurchaseListResponseDto.java
@@ -8,7 +8,7 @@ public record GroupPurchaseListResponseDto (
         String title,
         int price,
         int maxParticipants,
-        int currentParticipants,
+        Long currentParticipants,
         String progressStatus,
         String imageUrl
 ) {}
